[FIX]: React.Fragment를 사용해 unique key prop warning 해결
